fix(tui): Apply clippy auto-fixes - SPEC-957 Phase 2
Massive warning reduction via cargo clippy --fix --lib:
- 734 warnings → 97 warnings (87% reduction) ✅
- Exceeded <100 target

Auto-fixed patterns:
- Collapsible if statements (408 instances)
- Useless format! usage (32 instances)
- Unneeded return statements (24 instances)
- Let-binding unit values (38 instances)
- Map_or simplifications (13 instances)
- Push_str single-char literals (11 instances)
- Various style improvements (200+ instances)

Files modified: 94
Lines removed: 514 net

Remaining 97 warnings:
- 16: vec init then push
- 8: identical if blocks
- 7: collapsible if let
- 66: misc style/complexity issues

+++ codex-rs/tui/src/glitch_animation.rs
@@ -187,7 +187,7 @@ fn mask_to_outline_fill_lines(
             else if border[y][x] && xi <= reveal_x_outline.max(reveal_x_fill) {
                 let base = gradient_multi(x as f32 / (w.max(1) as f32));
                 // marching ants along diagonals
-                let period = (2 * scale_or(scale, 4)) as usize; // ~scale-based speed/size
+                let period = 2 * scale_or(scale, 4); // ~scale-based speed/size
                 let on = ((x + y + (frame as usize)) % period) < (period / 2);
                 let c = if on { bump_rgb(base, 0.22) } else { base };
                 // Make outline very light in final state
@@ -249,7 +249,7 @@ fn mask_to_outline_fill_lines_with_alpha(
             else if border[y][x] && xi <= reveal_x_outline.max(reveal_x_fill) {
                 let base = gradient_multi(x as f32 / (w.max(1) as f32));
                 // marching ants along diagonals
-                let period = (2 * scale_or(scale, 4)) as usize; // ~scale-based speed/size
+                let period = 2 * scale_or(scale, 4); // ~scale-based speed/size
                 let on = ((x + y + (frame as usize)) % period) < (period / 2);
                 let c = if on { bump_rgb(base, 0.22) } else { base };
                 // Make outline very light in final state
